+++ nextjs/app/fechahora/page.jsx
@@ -3,21 +3,21 @@ import React, { useState, useEffect } from "react";
 import Navbar from "../components/Navbar";
 import Calendar from "../components/Calendar";
 import Appointment from "../components/Appointment";
-import {token} from '../../public/token' 
+import { token } from '../../public/token'
 
 export default function DateReservation() {
     const [selectedDate, setSelectedDate] = useState(null);
     const [availableAppointments, setAvailableAppointments] = useState([]);
-    const[fechaHora, setFechaHora] = useState("");
-    const[servicio, setServicio] = useState(localStorage.getItem('tipoServicio'))
+    const [fechaHora, setFechaHora] = useState("");
+    const [servicio, setServicio] = useState(localStorage.getItem('tipoServicio'))
 
-    const[turno, SetTurno]= useState({
-        fechahora:"",
-        nota:"",
-        idUsuario:"1",
-        idServicio:"",
+    const [turno, SetTurno] = useState({
+        fechahora: "",
+        nota: "",
+        idUsuario: "1",
+        idServicio: "",
     });
-   
+
     useEffect(() => {
         if (selectedDate) {
             fetchAppointments(selectedDate);
@@ -66,29 +66,28 @@ export default function DateReservation() {
             hourArray.push(`${formattedHour}:00`);
         }
         return hourArray;
-    };  
-      const handleSubmit = async (e) => {
-        
+    };
+    const handleSubmit = async (e) => {
+
         e.preventDefault();
         console.log("paso")
         setFechaHora(localStorage.getItem('dia'));
         SetTurno({
             fechaHora: fechaHora,
-            nota:"",
-            idUsuario:"1",
+            nota: "",
+            idUsuario: "1",
             idServicio: servicio,
         })
-        
-         console.log(turno)
-       //Cookies.set('usuario', JSON.stringify(usuario));
-       //window.location.href = 'home'
-     };
+
+        console.log(turno)
+        //Cookies.set('usuario', JSON.stringify(usuario));
+        //window.location.href = 'home'
+    };
 
     const updatedHourArray = generateHourArray().filter(hour => !availableAppointments.some(appointment => clearString(appointment.fechahora) === hour));
 
     return (
         <div>
-            <Navbar></Navbar>
             <div class="h-screen max-w-full bg-[#303030] flex justify-around">
                 <div class=" pt-8 grid grid-cols-2 gap-x-14 ">
                     <div class="colspan-1">
@@ -109,17 +108,17 @@ export default function DateReservation() {
                             <div class="h-96 p-6 background rounded-xl overflow-y-scroll scroll-smooth">
                                 <div class="flex flex-col w-full">
                                     {updatedHourArray.map((updatedHour, index) => (
-                                        <Appointment key={index} hour={updatedHour}/>
+                                        <Appointment key={index} hour={updatedHour} />
                                     ))}
                                 </div>
                             </div>
                             <div class="grid grid-cols-2 gap-x-8 pt-8">
-                                <button onClick={(e) => {e.preventDefault(); window.location.href = '/'}}
-                                        class="colspan-1 bg-[#5865F2] hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-xl">
+                                <button onClick={(e) => { e.preventDefault(); window.location.href = '/' }}
+                                    class="colspan-1 bg-[#5865F2] hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-xl">
                                     Atrás
                                 </button>
-                                <button onClick={(e) => {e.preventDefault(); window.location.href = 'formulario'}}
-                                        class="colspan-1 bg-[#5865F2] hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-xl">
+                                <button onClick={(e) => { e.preventDefault(); window.location.href = 'formulario' }}
+                                    class="colspan-1 bg-[#5865F2] hover:bg-blue-700 text-white font-bold py-2 px-4 rounded-xl">
                                     Siguiente
                                 </button>
                             </div>
