+++ nextjs/next.config.mjs
@@ -1,11 +1,11 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    webpackDevMiddleware: config => {
-      config.watchOptions = {
-        poll: 1000,
-        aggregateTimeout: 300,
-      }
-      return config
-    },
+    // webpackDevMiddleware: config => {
+    //   config.watchOptions = {
+    //     poll: 1000,
+    //     aggregateTimeout: 300,
+    //   }
+    //   return config
+    // },
     };
 export default nextConfig;
