+++ nextjs/public/token.js
@@ -1,2 +1,2 @@
-const token="eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIyNzEwODQzMiIsImlhdCI6MTcxMTM2MTY1OSwiZXhwIjoxNzExMzYzMDk5fQ.m-diKiC5PHs-VkR_vIPwhJSa1HqJXV_mVsC9q4q34wo"
+const token="eyJhbGciOiJIUzI1NiJ9.eyJzdWIiOiIxMjM0NTY3OCIsImlhdCI6MTcxMTM5MDc4NSwiZXhwIjoxNzExMzkyMjI1fQ.JpkRrTqkgf7e60O7O9n-IXODdoSrBbBFKrsFkNAnzEA"
 module.exports={token}
